validation fix: removed check for sorting of time_first,time_last
Removed check for whether time_first and time_last series are sorted in an "irregular_intervals" timeseries. Since the intervals can be of different lengths, it's not possible to sort series by both time_first and time_last. Therefore, we doesn't require neither of them to be sorted (such requirement has never been stated in the specification).