Refreshing module configuration.
CallbackRegistrar offers a new way of registering hooks
`on_entity_creation`, which allows for refreshing the set values in
the next snapshot run.

The Control module offers a new endpoint for this purpose,
which calls the `load_config` of the appropriate module and sets the
flag to perform the refresh in snapshots.